Implement bomb placement and detonation, but not the flames
With a simple inflation animation to give feedback on the bombs
remaining time

+++ src/main/java/ecse321/fall2014/group3/bomberman/world/tile/timed/Fire.java
@@ -3,7 +3,6 @@ package ecse321.fall2014.group3.bomberman.world.tile.timed;
 import com.flowpowered.math.vector.Vector2f;
 
 import ecse321.fall2014.group3.bomberman.nterface.SpriteInfo;
-import ecse321.fall2014.group3.bomberman.world.tile.timed.TimedTile;
 
 public class Fire extends TimedTile {
     private static final SpriteInfo FIRE_SPRITE = new SpriteInfo("terrain", 1, Vector2f.ONE);
@@ -18,7 +17,7 @@ public class Fire extends TimedTile {
     }
 
     @Override
-    public int getRemainingTime() {
+    public long getRemainingTime() {
         return 0;
     }
 }
