+++ src/main/java/ecse321/fall2014/group3/bomberman/world/tile/timed/TimedTile.java
@@ -9,5 +9,5 @@ public abstract class TimedTile extends CollidableTile {
         super(position);
     }
 
-    public abstract int getRemainingTime();
+    public abstract long getRemainingTime();
 }
